% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R
\name{pet}
\alias{pet}
\title{pet}
\usage{
pet(name = NULL, animal = NULL)
}
\arguments{
\item{name}{A string containing the pet's name. Default: random name}

\item{animal}{A string containing the pet's species. Default: random species}
}
\value{
String output in the console introducing you to the pet
}
\description{
Get a new pet by finding out its name and species
Output: Prints string in console
}
\examples{
pet(name = "Old Yeller", animal = "dog")
}
