% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friend.R
\name{friend}
\alias{friend}
\title{friend}
\usage{
friend(name = NULL, job = NULL)
}
\arguments{
\item{name}{A string containing the friend's name. Picks a random name by default}

\item{job}{A string containing the friend's job. Picks a random job by default}
}
\value{
String output in the console introducing you to the friend
}
\description{
Make a new friend by getting to know their name and job (printed console output)
}
\examples{
pet(name = "David", job = "singer")
}
